Add all error calculation funtions to helper2.py, rewrite analysis.py example to visualize delay error and prediction error, implement simple prediction model that predicticts from the latest available point to time in the future by applying latest available velosity, rewrite plase visualization script
